// src/types/types.ts
export type ConfigType = {
  VIDEO_TITLE?: string;
  VIDEO_DESCRIPTION?: string;
  VIDEO_VISIBILITY?: 'PUBLIC' | 'UNLISTED' | 'PRIVATE';
  VIDEO_MADE_FOR_KIDS?: boolean;
  VIDEO_CATEGORY?: VideoCategoriesType;
  VIDEO_TAGS?: string[];
};

export enum VideoCategoriesType {
  Film = 'CREATOR_VIDEO_CATEGORY_FILM',
  Autos = 'CREATOR_VIDEO_CATEGORY_AUTOS',
  Music = 'CREATOR_VIDEO_CATEGORY_MUSIC',
  Pets = 'CREATOR_VIDEO_CATEGORY_PETS',
  Sports = 'CREATOR_VIDEO_CATEGORY_SPORTS',
  Travel = 'CREATOR_VIDEO_CATEGORY_TRAVEL',
  Gadgets = 'CREATOR_VIDEO_CATEGORY_GADGETS',
  People = 'CREATOR_VIDEO_CATEGORY_PEOPLE',
  Comedy = 'CREATOR_VIDEO_CATEGORY_COMEDY',
  Entertainment = 'CREATOR_VIDEO_CATEGORY_ENTERTAINMENT',
  News = 'CREATOR_VIDEO_CATEGORY_NEWS',
  HowTo = 'CREATOR_VIDEO_CATEGORY_HOWTO',
  Education = 'CREATOR_VIDEO_CATEGORY_EDUCATION',
  Science = 'CREATOR_VIDEO_CATEGORY_SCIENCE',
  Goverment = 'CREATOR_VIDEO_CATEGORY_GOVERNMENT',
}